Made the `weirdFunc1` function more efficient
The previous implementation would do 8 for loop iterations even when it wasn't needed, this was pretty stupid but that's what twitter's js code does.

All the code really needs is for the number to be divided by 2 until it's absolute value is < 1 because at that point it will never be an even number again (besides 0) therefore it can never add to the final number again.

I also changed how it decides to do a for loop because usually the code will only input 1 number that is (abs) > 1 anyways so doing math and checks on it is useless.